% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_funcs.R
\name{get_database_path}
\alias{get_database_path}
\title{Get the package database path}
\usage{
get_database_path()
}
\value{
The local path, in which the road.sqlite is located.
}
\description{
Get the package database path
}
