% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/road_package_aux.R
\name{road_get_archaeology}
\alias{road_get_archaeology}
\title{Get archaeology from ROAD database}
\usage{
road_get_archaeology(
  continents = NULL,
  subcontinents = NULL,
  countries = NULL,
  locality_types = NULL,
  cultural_periods = NULL,
  categories = NULL,
  age_min = NULL,
  age_max = NULL,
  archaeological_category = NULL,
  assemblages = NULL,
  localities = NULL
)
}
\arguments{
\item{continents}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{subcontinents}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{countries}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{locality_types}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{cultural_periods}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{categories}{string (one item) or vector of strings (one or more items).}

\item{age_min}{integer; minimum age of assemblage.}

\item{age_max}{integer; maximum age of assemblage.}

\item{archaeological_category}{string (one item) or vector of strings ('lithics', 'organic tools', 'symbolic artefacts', 'miscellaneous finds', 'feature');}

\item{assemblages}{list of assemblages; return value from function `road_get_assemblages`.}

\item{localities}{list of localities; return value from function `road_get_localities`.}
}
\value{
Database search result as list of archaeological finds.
}
\description{
`road_get_archaeology` fetches data of archaeological finds from ROAD database.
}
\details{
Archaeological finds are often part of an assemblage which means the function needs a list of
assemblages (return value of function `road_get_assemblages`) as its parameter.
}
\examples{
road_get_archaeology(continents = "Europe")
road_get_archaeology(continents = "Europe", archaeological_category = "feature")
road_get_archaeology(continents = "Europe", archaeological_category = c("feature", "symbolic artefacts"))
}
