% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_archaeology.R
\name{road_get_feature}
\alias{road_get_feature}
\title{Get feature assemblages from ROAD database}
\usage{
road_get_feature(
  continents = NULL,
  subcontinents = NULL,
  countries = NULL,
  locality_types = NULL,
  cultural_periods = NULL,
  categories = NULL,
  age_min = NULL,
  age_max = NULL,
  feature_interpretation = NULL,
  assemblages = NULL
)
}
\arguments{
\item{continents}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{subcontinents}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{countries}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{locality_types}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{cultural_periods}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{categories}{string (one item) or vector of strings (one or more items).}

\item{age_min}{integer; minimum age of assemblage.}

\item{age_max}{integer; maximum age of assemblage.}

\item{feature_interpretation}{string (one item) or vector of strings}

\item{assemblages}{list of assemblages; return value from function `road_get_assemblages`.}
}
\value{
Database search result as list of feature finds.
}
\description{
`road_get_feature` fetches data of feature finds from ROAD database.
}
\details{
Feature assemblages refer to archaeological features such as hearths, pits, or structures found at a site.
This function enables you to query feature data from the ROAD database using parameters like geographical location,
cultural periods, feature interpretation, and assemblages. Use the parameters to filter the results or omit them for broader results.
}
\examples{
road_get_feature(continents = "Europe")
road_get_feature(continents = "Europe", feature_interpretation = "bedding")
}
