% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_archaeology.R
\name{road_get_lithic_raw_material}
\alias{road_get_lithic_raw_material}
\title{Get lithic raw material from ROAD database}
\usage{
road_get_lithic_raw_material(
  continents = NULL,
  subcontinents = NULL,
  countries = NULL,
  locality_types = NULL,
  cultural_periods = NULL,
  categories = NULL,
  age_min = NULL,
  age_max = NULL,
  raw_material_list = NULL,
  assemblages = NULL
)
}
\arguments{
\item{continents}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{subcontinents}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{countries}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{locality_types}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{cultural_periods}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{categories}{string (one item) or vector of strings (one or more items).}

\item{age_min}{integer; minimum age of assemblage.}

\item{age_max}{integer; maximum age of assemblage.}

\item{raw_material_list}{string (one item) or vector of strings}

\item{assemblages}{list of assemblages; return value from function `road_get_assemblages`.}
}
\value{
Database search result as list of lithic finds with info about raw material.
}
\description{
`road_get_lithic_raw_material` fetches data of lithic finds from ROAD database.
}
\details{
Lithic raw material refers to the types of stone used for tool production in archaeological contexts.
This function allows you to query lithic raw material data from the ROAD database using parameters such as
geographical location, cultural periods, raw material types, and assemblages. Use the parameters to filter
the results or omit them to retrieve a broader dataset.
}
\examples{
road_get_lithic_raw_material(continents = "Europe")
road_get_lithic_raw_material(continents = "Europe", raw_material_list = "flint")
}
