% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_archaeology.R
\name{road_get_lithic_typology}
\alias{road_get_lithic_typology}
\title{Get lithic typology from ROAD database}
\usage{
road_get_lithic_typology(
  continents = NULL,
  subcontinents = NULL,
  countries = NULL,
  locality_types = NULL,
  cultural_periods = NULL,
  categories = NULL,
  age_min = NULL,
  age_max = NULL,
  tool_list = NULL,
  assemblages = NULL
)
}
\arguments{
\item{continents}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{subcontinents}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{countries}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{locality_types}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{cultural_periods}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{categories}{string (one item) or vector of strings (one or more items).}

\item{age_min}{integer; minimum age of assemblage.}

\item{age_max}{integer; maximum age of assemblage.}

\item{tool_list}{string (one item) or vector of strings}

\item{assemblages}{list of assemblages; return value from function `road_get_assemblages`.}
}
\value{
Database search result as list of lithic finds with info about typology.
}
\description{
`road_get_lithic_typology` fetches data of lithic finds from ROAD database.
}
\details{
Lithic typology refers to the classification of stone tools based on their shape, technology, and function.
This function enables you to query lithic typology data from the ROAD database using various parameters such as
geographical location, cultural periods, tool types, and assemblages. Use the parameters to filter the results
according to your research needs, or omit them to retrieve a broader dataset.
}
\examples{
road_get_lithic_typology(continents = "Europe")
road_get_lithic_typology(continents = "Europe", tool_list = "flake")
}
