% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/road_package_outdated_funcs.R
\name{road_get_paleofauna_}
\alias{road_get_paleofauna_}
\title{Get paleofauna finds from ROAD database}
\usage{
road_get_paleofauna_(
  continents = NULL,
  subcontinents = NULL,
  countries = NULL,
  locality_types = NULL,
  cultural_periods = NULL,
  categories = NULL,
  age_min = NULL,
  age_max = NULL,
  fauna_genus = NULL,
  fauna_species = NULL,
  assemblages = NULL,
  localities = NULL
)
}
\arguments{
\item{continents}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{subcontinents}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{countries}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{locality_types}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{cultural_periods}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{categories}{string (one item) or vector of strings (one or more items).}

\item{age_min}{integer; minimum age of assemblage.}

\item{age_max}{integer; maximum age of assemblage.}

\item{fauna_genus}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{fauna_species}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{assemblages}{list of assemblages; return value from function `road_get_assemblages`.}

\item{localities}{list of localities; return value from function `road_get_localities`.}
}
\value{
Database search result as list of archaeological finds.
}
\description{
`road_get_paleofauna_` fetches data of paleofauna finds from ROAD database.
}
\details{
Paleofauna finds are often part of an assemblage which means the function needs a list of
assemblages (return value of function `road_get_assemblages`) as its parameter.
}
\examples{
road_get_paleofauna(continents = "Europe")
road_get_paleofauna(continents = "Europe", archaeological_category = "feature")
road_get_paleofauna(continents = "Europe", archaeological_category = c("feature", "symbolic artefacts"))
}
