% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{road_list_parameter_values}
\alias{road_list_parameter_values}
\title{Get parameter value from ROAD Database}
\usage{
road_list_parameter_values(function_name, function_parameter)
}
\arguments{
\item{function_name}{name of a function}

\item{function_parameter}{name of a function parameter}
}
\value{
List of attribute values.
}
\description{
`road_list_parameter_values` fetches values of a given parameter in the database or 
read values from file
}
