% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{road_list_values_}
\alias{road_list_values_}
\title{Get attribute value from ROAD Database}
\usage{
road_list_values_(attribute_name)
}
\arguments{
\item{attribute_name}{name of an attribute.}
}
\value{
List of attribute values.
}
\description{
`road_list_values_` fetches values of a given attribute in the database or 
read values from file
}
