% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_funcs.R
\name{road_run_query}
\alias{road_run_query}
\title{unknown' is a correct value of 'transport_distance', we dont want replace it.}
\usage{
road_run_query(query)
}
\description{
unknown' is a correct value of 'transport_distance', we dont want replace it.
}
