% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{road_summerize_archaeology}
\alias{road_summerize_archaeology}
\title{Get table and attribute overview of ROAD database}
\usage{
road_summerize_archaeology(term = NULL)
}
\arguments{
\item{term}{string (one item)}
}
\value{
Database search result as list of lithic finds with info about typology.
}
\description{
`road_summerize_archaeology` fetches archaeological tables and their attributes from ROAD database.
}
\examples{
road_summerize_archaeology(term = "Cores")
}
