% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assemblages.R
\name{road_get_assemblages}
\alias{road_get_assemblages}
\title{Get assemblages from ROAD database}
\usage{
road_get_assemblages(
  continents = NULL,
  subcontinents = NULL,
  countries = NULL,
  locality_types = NULL,
  cultural_periods = NULL,
  categories = NULL,
  age_min = NULL,
  age_max = NULL
)
}
\arguments{
\item{continents}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{subcontinents}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{countries}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{locality_types}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{cultural_periods}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{categories}{string (one item) or vector of strings (one or more items).}

\item{age_min}{integer; minimum age of assemblage.}

\item{age_max}{integer; maximum age of assemblage.}
}
\value{
Database search result as list of assemblages.
}
\description{
`road_get_assemblages` fetches data of archeological assemblages from ROAD database. Assembalges are articulated archeological finds inside in a locality. One locality
can host multiple assemblages which can for example be associated with certain
geological layers or historical time periods.
}
\examples{
road_get_assemblages(countries = c("Germany", "France"), age_min = 100000L)
road_get_assemblages(categories = "human remains", age_max = 100000L)
road_get_assemblages(subcontinents = "Central Asia", cultural_periods = "Middle Paleolithic")
}
