% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assemblages.R
\name{road_get_assemblages}
\alias{road_get_assemblages}
\title{Get assemblages from ROAD database}
\usage{
road_get_assemblages(
  continent = NULL,
  subcontinent = NULL,
  country = NULL,
  locality_type = NULL,
  cultural_period = NULL,
  technocomplex = NULL,
  category = NULL,
  age_min = NULL,
  age_max = NULL
)
}
\arguments{
\item{continent}{specifies the continent(s) (e.g. Africa, Europe, Asia).
Run \code{road_list_argument_values("continent")} to display possible values.
The argument \code{continent} is a string (one item) or vector of strings
(one or more items); defaults to NULL.}

\item{subcontinent}{specifies the continental region(s) (e.g. Southern Europe).
Run \code{road_list_argument_values("subcontinent")} to display possible values.
The argument \code{subcontinent} is a string (one item) or vector of strings
(one or more items); defaults to NULL.}

\item{country}{specifies the name of the country where a locality is situated
(e.g. Germany, Kenya, Saudi Arabia). Run \code{road_list_argument_values("country")}
to display possible values.
The argument \code{country} is a string (one item) or vector of strings
(one or more items); defaults to NULL.}

\item{locality_type}{specifies the type of locality (e.g. cave, rockshelter, open air).
Run \code{road_list_argument_values("locality_type")} to display possible values.
The argument \code{locality_type} is a string (one item) or vector of strings
(one or more items); defaults to NULL.}

\item{cultural_period}{specifies the main cultural epoch(s) and includes the
Eurasian Paleolithic (Lower, Middle, Upper, Epi-) and the African Stone Age
(Earlier, Middle, Later). Run \code{road_list_argument_values("cultural_period")}
to display possible values. The argument \code{cultural_period} is a string
(one item) or vector of strings (one or more items); defaults to NULL.}

\item{technocomplex}{specifies an archaeological culture or named stone tool
industry (e.g. Oldowan, Acheulean, Mousterian).
Run \code{road_list_argument_values("technocomplex")} to display possible values.
The argument \code{technocomplex} is a string (one item) or vector of strings
(one or more items); defaults to NULL.}

\item{category}{specifies the assemblage category with the classes
human remains, raw material, typology, technology, function, organic tools,
symbolic artifacts, feature, miscellaneous finds, paleofauna, animal remains,
plant remains. The argument \code{category} is a string (one item) or
vector of strings (one or more items); defaults to NULL.}

\item{age_min}{specifies the minimum age in years before present, using 1950 CE
as the baseline. If possible the argument \code{age_min} will be converted to an integer; defaults to NULL.}

\item{age_max}{specifies the maximum age in years before present, using 1950 CE
as the baseline. If possible the argument \code{age_max} will be converted to an integer; defaults to NULL.}
}
\value{
Database search result as a data frame with the information about
assemblages like their geographic information, cultural period, locality type,
assemblage category, dating and quality of collection.
}
\description{
The \strong{\code{road_get_assemblages}} fetches data of assemblages from ROAD database.
The ROAD table assemblage contains information about classes of finds. An assemblage is
defined as a collected find consisting of grouped classes of materials, for example:
archaeological finds (including raw material, typology, technology, function, organic tools,
symbolic artifacts, feature and miscellaneous finds), human remains, faunal remains or
botanical remains. An assemblage comes from a single geological layer or several
geological layers of a locality, whether the nature of the locality is geological,
archaeological or paleontological. Each physical object contained in any assemblage
can appear only once in this table.
}
\details{
Use parameters to filter search results by location, type, culture, assemblage
category, age or omit them to have a broader result set.
All parameters are optional and should be omitted or set to NULL when not used.
}
\examples{
road_get_assemblages(country = c("Germany", "France"), age_min = 300000L,
                     category = c("miscellaneous finds"))
road_get_assemblages(subcontinent = c("Caucasus"), category = "human remains",
                     age_max = 100000L)
}
