% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_assemblages.R
\name{road_get_assemblages}
\alias{road_get_assemblages}
\title{Get assemblages from ROAD database}
\usage{
road_get_assemblages(
  continents = NULL,
  subcontinents = NULL,
  countries = NULL,
  locality_types = NULL,
  cultural_periods = NULL,
  categories = NULL,
  age_min = NULL,
  age_max = NULL
)
}
\arguments{
\item{continents}{specifies the continent(s) of the country/countries, e.g. Africa, Europe, Asia. The parameter continents is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{subcontinents}{specifies the continental region(s) of the country , e.g. Southern Europe. The parameter subcontinents is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{countries}{specifies the name of the country where a locality is situated (e.g.
Germany, Kenya, Saudi Arabia, China). The parameter countries is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{locality_types}{specifies the type of locality (e.g. cave, rockshelter, open air, profile, outcrop,
mine, quarry, boring). The parameter locality_types is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{cultural_periods}{specifies the main cultural epoch(s) and includes the Eurasian
Paleolithic (Lower, Middle, Upper, Epi) and the African Stone Age (Earlier, Middle, Later). The parameter cultural_periods is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{categories}{specifies the assemblage category/categories. For an assemblage consisting of human remains category includes the entry “human
remains”. In the case of archaeological assemblages, multiple categories are the norm and may
include “raw material, typology, technology, function, organic tools, symbolic artifacts, feature,
miscellaneous finds”. A faunal assemblage can also contain multiple entries including
“paleofauna, animal remains”, while a botanical assemblage can only include the entry “plant
remains”.The parameter categories is a string (one item) or vector of strings (one or more items).}

\item{age_min}{specifies the minimum age of assemblage. The parameter age_min is an integer.}

\item{age_max}{specifies the maximum age of assemblage. The parameter age_max is an integer.}
}
\value{
Database search result as a data frame with the information about 
assemblages like their geographic information, cultural period, locality type, 
assemblage category and dating.
}
\description{
The \strong{\code{road_get_assemblages}} fetches data of assemblages from ROAD database. The ROAD table assemblage contains information about classes of finds. An assemblage is
defined as a collected find consisting of grouped classes of materials, for example:
archaeological finds (including raw material, typology, technology, function, organic tools,
symbolic artifacts, feature and miscellaneous finds), human remains, faunal remains or
botanical remains. An assemblage comes from a single geological layer or several geological layers of a locality, whether
the nature of the locality is geological, archaeological or paleontological. Each physical object
contained in any assemblage can appear only once in this table.
}
\details{
Use parameters to filter search results by location, type, culture, assemblage category, age or omit them to have a broader result set.
All parameters are optional and should be omitted or set to NULL when not used.
}
\examples{
road_get_assemblages(countries = c("Germany", "France"), age_min = 300000L, 
                     categories = c("miscellaneous finds"))
road_get_assemblages(subcontinents = c("Caucasus"), categories = "human remains", 
                     age_max = 100000L)
}
