% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dates.R
\name{road_get_dates}
\alias{road_get_dates}
\title{Retrieve Dating Information for Assemblages, Geolayers, and Archlayers from the ROAD Database}
\usage{
road_get_dates(
  continents = NULL,
  subcontinents = NULL,
  countries = NULL,
  locality_types = NULL,
  cultural_periods = NULL,
  dating_methods = NULL,
  material_dated = NULL,
  age_min = NULL,
  age_max = NULL,
  technocomplex = NULL,
  categories = NULL,
  assemblages = NULL
)
}
\arguments{
\item{continents}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{subcontinents}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{countries}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{locality_types}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{cultural_periods}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{dating_methods}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{material_dated}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{age_min}{integer; defaults to NULL.}

\item{age_max}{integer; defaults to NULL.}

\item{technocomplex}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{categories}{string (one item) or vector of strings (one or more items).}

\item{assemblages}{list of assemblages; return value from function `road_get_assemblages`.}
}
\value{
A data frame with absolute dating results. Rows represent dates, columns contain details about minimum and maximum ages, dating method, and material.
}
\description{
The `road_get_dates` function retrieves absolute dating records for assemblages, geological layers, and archaeological layers from the ROAD database.
}
\details{
The function is designed for users analyzing time series and allows for 
filtering results based on various criteria.

Background on Dating Context:
Not all finds can be directly dated, either due to potential damage during
sampling or because the materials are unsuitable for absolute dating. In such
cases, the surrounding sediment or other objects—made from more suitable
materials—are dated and associated with the find. As a result, absolute dates
are stored in three different tables within ROAD: 
* `geological_layer_age`: Dates from samples taken from geological layers.
* `archaeological_layer_age`: Dates from samples taken from archaeological layers.
* `assemblage_age`: Dates from samples taken directly from objects.

Background on Multiple Dates: Multiple dates may be available for a single object. For example, both the top and bottom of a stratigraphic unit may be sampled to bracket the age of a find. Different dating methods might also be used to obtain robust results, which can lead to varying age estimates. This is especially true when earlier dating methods are updated with more modern techniques. It is recommended to cross-check all results against relevant literature. 

Background on Aggregated Ages: The ROCEEH project has developed its own model to aggregate both absolute and relative dating methods to derive the variables `age_min` and `age_max`. These aggregated ages are not included in the `road_get_dates` function, which only returns raw data.

Background on Radiocarbon Dating: Please note that C14 dates are uncalibrated. For further analysis, use one of the available calibration tools, such as the `rcarbon` package available on CRAN (see: https://cran.r-project.org/web/packages/rcarbon/index.html).

All Parameters are optional. If not used, omit them or set them to `NULL`.
}
\examples{
road_get_dates(dating_methods = c("geology", "biostratigraphy"))
road_get_dates(material_dated = c("coprolite", "glass", "ivory"), age_min = 10000L, 
                         age_max = 100000L, dating_methods = c("geology", "biostratigraphy"))
road_get_dates(countries = c("Germany", "France"), cultural_periods = "Middle Paleolithic")                         
}
