% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dates.R
\name{road_get_dates}
\alias{road_get_dates}
\title{Retrieve Dating Information for Assemblages, Geolayers, and Archlayers from the ROAD Database}
\usage{
road_get_dates(
  continents = NULL,
  subcontinents = NULL,
  countries = NULL,
  locality_types = NULL,
  cultural_periods = NULL,
  technocomplexes = NULL,
  dating_methods = NULL,
  material_dated = NULL,
  age_min = NULL,
  age_max = NULL,
  categories = NULL,
  assemblages = NULL
)
}
\arguments{
\item{continents}{specifies the continent(s) (e.g. Africa, Europe, Asia).
Run \code{road_list_argument_values("continents")} to display possible values.
The argument \code{continents} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{subcontinents}{specifies the continental region(s) (e.g. Southern Europe). 
Run \code{road_list_argument_values("subcontinents")} to display possible values.
The argument \code{subcontinents} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{countries}{specifies the name of the country where a locality is situated 
(e.g. Germany, Kenya, Saudi Arabia). Run \code{road_list_argument_values("countries")} 
to display possible values.
The argument \code{countries} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{locality_types}{specifies the type of locality (e.g. cave, rockshelter, open air).
Run \code{road_list_argument_values("locality_types")} to display possible values.
The argument \code{locality_types} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{cultural_periods}{specifies the main cultural epoch(s) and includes the 
Eurasian Paleolithic (Lower, Middle, Upper, Epi-) and the African Stone Age 
(Earlier, Middle, Later). Run \code{road_list_argument_values("cultural_periods")} 
to display possible values. The argument \code{cultural_periods} is a string 
(one item) or vector of strings (one or more items); defaults to NULL.}

\item{technocomplexes}{specifies an archaeological culture or named stone tool 
industry (e.g. Oldowan, Acheulean, Mousterian).
Run \code{road_list_argument_values("technocomplexes")} to display possible values.
The argument \code{technocomplexes} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{dating_methods}{specifies the method of temporal analysis (e.g. 14C, OSL, IRSL).
Run \code{road_list_argument_values("dating_methods")} to display possible values.
The argument \code{dating_methods} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{material_dated}{(string) specifies the material used for temporal analysis 
(e.g. bone, tooth, antler). Run \code{road_list_argument_values("material_dated")} 
to display possible values. The argument \code{datied_material} is a string 
(one item) or vector of strings (one or more items); defaults to NULL.}

\item{age_min}{specifies the minimum age in years before present, using 1950 CE 
as the baseline. The argument \code{age_min} is an integer; defaults to NULL.}

\item{age_max}{specifies the maximum age in years before present, using 1950 CE 
as the baseline. The argument \code{age_max} is an integer; defaults to NULL.}

\item{categories}{specifies the assemblage category/categories with the classes 
human remains, raw material, typology, technology, function, organic tools, 
symbolic artifacts, feature, miscellaneous finds, paleofauna, animal remains, 
plant remains. The argument \code{categories} is a string (one item) or 
vector of strings (one or more items); defaults to NULL.}

\item{assemblages}{list of assemblages; return value from function 
\code{road_get_assemblages}. Can be used instead of the other locality and 
assemblage parameters to filter the results.}
}
\value{
A data frame with absolute dating information. Rows represent individual dates, columns contain standard outputs and dating-related details on:

\code{age}: The attribute specifies the result of the dating analysis as reported by the laboratory in years before present (BP). Note that 14C dates are \strong{uncalibarated}.

\code{negative_standard_deviation} & \code{positive_standard_deviation}: The attributes specify the positive and negative standard deviation of the dating analysis in years.

\code{material_dated}: The attribute specifies the general type of material analyzed (e.g. bone, tooth, antler etc.).

\code{dating_method}: The attribute specifies the method of analysis (e.g. 14C, OSL, IRSL, etc.).

\code{laboratory_idlaboratory}: The attribute is the official abbreviation for the designated analytical laboratory.
}
\description{
The  \strong{\code{road_get_dates}} function retrieves absolute dating records for 
assemblages, geological layers, and archaeological layers from the ROAD database.
}
\details{
The function is designed for users analyzing time series and allows for 
filtering results based on various criteria.

\strong{Background on Dating Context:}
Not all finds can be directly dated, either due to potential damage during
sampling or because the materials are unsuitable for absolute dating. In such
cases, the surrounding sediment or other objects—made from more suitable
materials—are dated and associated with the find. As a result, absolute dates
are stored in three different tables within ROAD:
\itemize{
   \item \code{geological_layer_age}: Dates from samples taken from geological layers.
   \item \code{archaeological_layer_age}: Dates from samples taken from archaeological layers.
   \item \code{assemblage_age}: Dates from samples taken directly from objects.
   }

\strong{Background on multiple dates:} Multiple dates may be available for a
single object. For example, both the top and bottom of a stratigraphic unit 
may be sampled to bracket the age of a find. Different dating methods might 
also be used to obtain robust results, which can lead to varying age estimates. 
This is especially true when earlier dating methods are updated with more 
modern techniques. It is recommended to cross-check all results against 
relevant literature. 

\strong{Background on aggregated ages:} The ROCEEH project has developed its 
own model to aggregate both absolute and relative dating methods to derive 
the variables \code{age_min} and \code{age_max}. These aggregated ages are 
not included in the \code{road_get_dates} function, which only returns raw data.

\strong{Background on radiocarbon dating:} Please note that C14 dates are 
uncalibrated. For further analysis, use one of the available calibration 
tools, such as the \pkg{rcarbon} package available on CRAN 
(see: \url{https://cran.r-project.org/web/packages/rcarbon/index.html}).

All parameters are optional. If not used, omit them or set them to NULL.
}
\examples{
road_get_dates(dating_methods = c("14C (radiocarbon) dating", "U series (uranium-thorium) dating"))
road_get_dates(material_dated = c("coprolite", "ivory"), 
               age_min = 10000L, 
               age_max = 50000L, 
               dating_method = "14C (radiocarbon) dating")
road_get_dates(countries = c("Germany", "France"), 
               cultural_periods = "Middle Paleolithic")                         
}
