% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/road_package_aux.R
\name{road_get_dates}
\alias{road_get_dates}
\title{Get dates for assemblages, geolayers and archlayers from ROAD Database}
\usage{
road_get_dates(
  continents = NULL,
  subcontinents = NULL,
  countries = NULL,
  locality_types = NULL,
  cultural_periods = NULL,
  dating_methods = NULL,
  material_dated = NULL,
  age_min = NULL,
  age_max = NULL,
  technocomplex = NULL,
  categories = NULL,
  assemblages = NULL,
  localities = NULL
)
}
\arguments{
\item{continents}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{subcontinents}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{countries}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{locality_types}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{cultural_periods}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{dating_methods}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{material_dated}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{age_min}{integer; defaults to NULL.}

\item{age_max}{integer; defaults to NULL.}

\item{technocomplex}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{categories}{string (one item) or vector of strings (one or more items).}

\item{assemblages}{list of assemblages; return value from function `road_get_assemblages`.}

\item{localities}{list of localities; return value from function `road_get_localities`.}
}
\value{
date records
}
\description{
`road_get_dates` fetches records of the age tables in ROAD
}
\details{
All parameters are optional and should be omitted or set to NULL when not used.
}
\examples{
road_get_dates(dating_methods = c("geology", "biostratigraphy"))
road_get_dates(material_dated = c("coprolite", "glass", "ivory"), age_min = 10000L, 
                         age_max = 100000L, dating_methods = c("geology", "biostratigraphy"))
road_get_dates(countries = c("Germany", "France"), cultural_periods = "Middle Paleolithic")                         
                         
                         
}
