% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/road_package_aux.R
\name{road_get_dates}
\alias{road_get_dates}
\title{Get dates for assemblages, geolayers and archlayers from ROAD Database}
\usage{
road_get_dates(
  dating_methods = NULL,
  material_dated = NULL,
  age_min = NULL,
  age_max = NULL
)
}
\arguments{
\item{dating_methods;}{defaults to NULL.}

\item{material_dated;}{defaults to NULL.}

\item{age_min;}{defaults to NULL.}

\item{age_max;}{defaults to NULL.}
}
\value{
date records
}
\description{
`road_get_dates` fetches records of the age tables in ROAD
}
\details{
All parameters are optional and should be omitted or set to NULL when not used.
}
\examples{
road_get_dates(dating_methods = c("geology", "biostratigraphy"))
road_get_dates(material_dated = c("coprolite", "glass", "ivory"), age_min = 10000L, 
                         age_max = 100000L, dating_methods = c("geology", "biostratigraphy"))
}
