% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_archaeology.R
\name{road_get_feature}
\alias{road_get_feature}
\title{Get feature assemblages from ROAD database}
\usage{
road_get_feature(
  continents = NULL,
  subcontinents = NULL,
  countries = NULL,
  locality_types = NULL,
  cultural_periods = NULL,
  categories = NULL,
  age_min = NULL,
  age_max = NULL,
  feature_interpretation = NULL,
  assemblages = NULL
)
}
\arguments{
\item{continents}{specifies the continent(s) of the country/countries, e.g. Africa, Europe, Asia. The parameter continents is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{subcontinents}{specifies the continental region(s) of the country , e.g. Southern Europe. The parameter subcontinents is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{countries}{specifies the name of the country where a locality is situated (e.g.
Germany, Kenya, Saudi Arabia, China). The parameter countries is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{locality_types}{specifies the type of locality (e.g. cave, rockshelter, open air, profile, outcrop,
mine, quarry, boring). The parameter locality_types is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{cultural_periods}{specifies the main cultural epoch(s) and includes the Eurasian
Paleolithic (Lower, Middle, Upper, Epi) and the African Stone Age (Earlier, Middle, Later). The parameter cultural_periods is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{categories}{specifies the assemblage category/categories. For an assemblage consisting of human remains category includes the entry “human
remains”. In the case of archaeological assemblages, multiple categories are the norm and may
include “raw material, typology, technology, function, organic tools, symbolic artifacts, feature,
miscellaneous finds”. A faunal assemblage can also contain multiple entries including
“paleofauna, animal remains”, while a botanical assemblage can only include the entry “plant
remains”.The parameter categories is a string (one item) or vector of strings (one or more items).}

\item{age_min}{specifies the minimum age of assemblage. The parameter age_min is an integer.}

\item{age_max}{specifies the maximum age of assemblage. The parameter age_max is an integer.}

\item{feature_interpretation}{specifies feature present in the archaeological assemblage. 
Feature can be: bedding, burial, butchering event, combustion feature, cupule, dumping area,
dwelling, impact feature, mine, pit, post hole, shell midden, stone construction, 
windbreak, etc. The parameter feature_interpretation is a string (one item) or vector of strings.}

\item{assemblages}{list of assemblages; return value from function `road_get_assemblages`.}
}
\value{
Database search result as list of feature finds with their geographic information, 
cultural period, locality type, category, dating and interpretations.
}
\description{
`road_get_feature` fetches data of feature finds from ROAD database. Feature assemblages refer to archaeological features such as hearths, pits, or structures found at a site.
This function enables you to query feature data from the ROAD database using parameters like geographical location,
cultural periods, feature interpretation, and assemblages. Use the parameters to filter the results or omit them for broader results.
}
\examples{
road_get_feature(continents = c("Europe", "Africa"), feature_interpretation = "textile imprints")
road_get_feature(continents = "Africa", locality_types = c("cave"), 
                 feature_interpretation = "bedding")
}
