% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_archaeology.R
\name{road_get_feature}
\alias{road_get_feature}
\title{Get feature assemblages from ROAD database}
\usage{
road_get_feature(
  continents = NULL,
  subcontinents = NULL,
  countries = NULL,
  locality_types = NULL,
  cultural_periods = NULL,
  technocomplexes = NULL,
  categories = NULL,
  age_min = NULL,
  age_max = NULL,
  feature_interpretation = NULL,
  assemblages = NULL
)
}
\arguments{
\item{continents}{specifies the continent(s) (e.g. Africa, Europe, Asia).
Run \code{road_list_argument_values("continents")} to display possible values.
The argument \code{continents} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{subcontinents}{specifies the continental region(s) (e.g. Southern Europe). 
Run \code{road_list_argument_values("subcontinents")} to display possible values.
The argument \code{subcontinents} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{countries}{specifies the name of the country where a locality is situated 
(e.g. Germany, Kenya, Saudi Arabia). Run \code{road_list_argument_values("countries")} 
to display possible values.
The argument \code{countries} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{locality_types}{specifies the type of locality (e.g. cave, rockshelter, open air).
Run \code{road_list_argument_values("locality_types")} to display possible values.
The argument \code{locality_types} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{cultural_periods}{specifies the main cultural epoch(s) and includes the 
Eurasian Paleolithic (Lower, Middle, Upper, Epi-) and the African Stone Age 
(Earlier, Middle, Later). Run \code{road_list_argument_values("cultural_periods")} 
to display possible values. The argument \code{cultural_periods} is a string 
(one item) or vector of strings (one or more items); defaults to NULL.}

\item{technocomplexes}{specifies an archaeological culture or named stone tool 
industry (e.g. Oldowan, Acheulean, Mousterian).
Run \code{road_list_argument_values("technocomplexes")} to display possible values.
The argument \code{technocomplexes} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{categories}{specifies the assemblage category/categories with the classes 
human remains, raw material, typology, technology, function, organic tools, 
symbolic artifacts, feature, miscellaneous finds, paleofauna, animal remains, 
plant remains. The argument \code{categories} is a string (one item) or 
vector of strings (one or more items); defaults to NULL.}

\item{age_min}{specifies the minimum age in years before present, using 1950 CE
as the baseline. If possible the argument \code{age_min} will be converted to an integer; defaults to NULL.}

\item{age_max}{specifies the maximum age in years before present, using 1950 CE
as the baseline. If possible the argument \code{age_max} will be converted to an integer; defaults to NULL.}

\item{feature_interpretation}{specifies archaeological features present in 
the archaeological assemblage (e.g. bedding, burial, butchering event).
Run \code{road_list_argument_values("feature_interpretation")} to display 
possible values. The argument \code{feature_interpretation} is a string 
(one item) or vector of strings; defaults to NULL.}

\item{assemblages}{specifies a data frame necessarily containing columns 
locality_id, assemblage_id. It can be  generated as return value of the 
function 'road_get_assemblages'. It can be used instead of the locality 
and assemblage search parameters to filter the results.}
}
\value{
Database search result as a data frame with the information about 
archaeological finds. Rows represent archaeological finds and contain standard 
information like locality name, geographic information, cultural period, 
locality type, assemblage category, dating and feature-related details:

\code{interpretations}: Interpretation of the feature present in the 
archaeological assemblage. Interpretations can be bedding, burial, butchering 
event, combustion feature, cupule, dumping area.
}
\description{
The \strong{\code{road_get_feature}} fetches data of feature finds from ROAD database. Feature assemblages refer to archaeological features such as hearths, pits, or structures found at a site.
This function enables you to query feature data from the ROAD database using parameters like geographical location,
cultural periods, feature interpretation, and assemblages. Use the parameters to filter the results or omit them for broader results.
}
\examples{
road_get_feature(continents = c("Europe", "Africa"), feature_interpretation = "textile imprints")
road_get_feature(continents = "Africa", locality_types = c("cave"),
                 feature_interpretation = "bedding")
}
