% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_archaeology.R
\name{road_get_lithic_raw_material}
\alias{road_get_lithic_raw_material}
\title{Get lithic raw material from ROAD database}
\usage{
road_get_lithic_raw_material(
  continents = NULL,
  subcontinents = NULL,
  countries = NULL,
  locality_types = NULL,
  cultural_periods = NULL,
  categories = NULL,
  age_min = NULL,
  age_max = NULL,
  raw_material_list = NULL,
  assemblages = NULL
)
}
\arguments{
\item{continents}{specifies the continent(s) of the country/countries, e.g. Africa, Europe, Asia. The parameter continents is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{subcontinents}{specifies the continental region(s) of the country , e.g. Southern Europe. The parameter subcontinents is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{countries}{specifies the name of the country where a locality is situated (e.g.
Germany, Kenya, Saudi Arabia, China). The parameter countries is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{locality_types}{specifies the type of locality (e.g. cave, rockshelter, open air, profile, outcrop,
mine, quarry, boring). The parameter locality_types is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{cultural_periods}{specifies the main cultural epoch(s) and includes the Eurasian
Paleolithic (Lower, Middle, Upper, Epi) and the African Stone Age (Earlier, Middle, Later). The parameter cultural_periods is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{categories}{specifies the assemblage category/categories. For an assemblage consisting of human remains category includes the entry “human
remains”. In the case of archaeological assemblages, multiple categories are the norm and may
include “raw material, typology, technology, function, organic tools, symbolic artifacts, feature,
miscellaneous finds”. A faunal assemblage can also contain multiple entries including
“paleofauna, animal remains”, while a botanical assemblage can only include the entry “plant
remains”.The parameter categories is a string (one item) or vector of strings (one or more items).}

\item{age_min}{specifies the minimum age of assemblage. The parameter age_min is an integer.}

\item{age_max}{specifies the maximum age of assemblage. The parameter age_max is an integer.}

\item{raw_material_list}{specifies raw materials like quartz, quartzite, chert, 
chert jurassic, chert cretaceous, tabular flint, flint baltic, flint tertiary, 
radiolarite, basalt, granite, obsidian, unknown, etc.
The parameter raw_material_list is a string (one item) or vector of strings}

\item{assemblages}{list of assemblages; return value from function `road_get_assemblages`.}
}
\value{
Database search result as list of lithic finds with their geographic information, cultural period, locality type, category, dating and info about raw material.
}
\description{
`road_get_lithic_raw_material` fetches data of lithic finds from ROAD database. 
Lithic raw material refers to the types of stone used for tool production in 
archaeological contexts. This function allows you to query lithic raw material 
data from the ROAD database using parameters such as geographical location, 
cultural periods, raw material types, and assemblages. Use the parameters to filter
the results or omit them to retrieve a broader dataset.
}
\examples{
road_get_lithic_raw_material(continents = "Europe")
road_get_lithic_raw_material(continents = "Europe", raw_material_list = "flint")
}
