% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_archaeology.R
\name{road_get_lithic_typology}
\alias{road_get_lithic_typology}
\title{Get lithic typology from ROAD database}
\usage{
road_get_lithic_typology(
  continents = NULL,
  subcontinents = NULL,
  countries = NULL,
  locality_types = NULL,
  cultural_periods = NULL,
  categories = NULL,
  age_min = NULL,
  age_max = NULL,
  tool_list = NULL,
  assemblages = NULL
)
}
\arguments{
\item{continents}{specifies the continent(s) of the country/countries, e.g. Africa, Europe, Asia. The parameter continents is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{subcontinents}{specifies the continental region(s) of the country , e.g. Southern Europe. The parameter subcontinents is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{countries}{specifies the name of the country where a locality is situated (e.g.
Germany, Kenya, Saudi Arabia, China). The parameter countries is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{locality_types}{specifies the type of locality (e.g. cave, rockshelter, open air, profile, outcrop,
mine, quarry, boring). The parameter locality_types is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{cultural_periods}{specifies the main cultural epoch(s) and includes the Eurasian
Paleolithic (Lower, Middle, Upper, Epi) and the African Stone Age (Earlier, Middle, Later). The parameter cultural_periods is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{categories}{specifies the assemblage category/categories. For an assemblage consisting of human remains category includes the entry “human
remains”. In the case of archaeological assemblages, multiple categories are the norm and may
include “raw material, typology, technology, function, organic tools, symbolic artifacts, feature,
miscellaneous finds”. A faunal assemblage can also contain multiple entries including
“paleofauna, animal remains”, while a botanical assemblage can only include the entry “plant
remains”.The parameter categories is a string (one item) or vector of strings (one or more items).}

\item{age_min}{specifies the minimum age of assemblage. The parameter age_min is an integer.}

\item{age_max}{specifies the maximum age of assemblage. The parameter age_max is an integer.}

\item{tool_list}{specifies values that can be entered for various tool types. Tool types
can contain 1) chipped tool types like scraper end, scraper side, scraper carinated, burin, handaxe, chopper,
cleaver, point, point unifacial, segment, unknown; 2) non-chipped tool types 
like grindstone upper, hammerstone, anvil, retoucher; 3) non-tools like core, debitage, 
flake, point; 4) unknown like cobble, block, manuport. 
The parameter tool_list is a string (one item) or vector of strings}

\item{assemblages}{list of assemblages; return value from function `road_get_assemblages`.}
}
\value{
Database search result as list of lithic finds with their geographic 
information, cultural period, locality type, category, dating and info about typology.
}
\description{
`road_get_lithic_typology` fetches data of lithic finds from ROAD database.
Lithic typology refers to the classification of stone tools based on their shape, technology, and function.
This function enables you to query lithic typology data from the ROAD database using various parameters such as
geographical location, cultural periods, tool types, and assemblages. Use the parameters to filter the results
according to your research needs, or omit them to retrieve a broader dataset.
}
\examples{
road_get_lithic_typology(continents = "Europe")
road_get_lithic_typology(continents = "Europe", tool_list = "flake")
}
