% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_localities.R
\name{road_get_localities}
\alias{road_get_localities}
\title{Get localities from ROAD Database}
\usage{
road_get_localities(
  continents = NULL,
  subcontinents = NULL,
  countries = NULL,
  locality_types = NULL,
  cultural_periods = NULL
)
}
\arguments{
\item{continents}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{subcontinents}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{countries}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{locality_types}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{cultural_periods}{string (one item) or vector of strings (one or more items); defaults to NULL.}
}
\value{
Database search result as list of localities.
}
\description{
`road_get_localities` fetches data of archaeological sites (localities) from ROAD database.
}
\details{
Use parameters to spatially delimit search results or omit them to have a broader radius.
All parameters are optional and should be omitted or set to NULL when not used.
}
