% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_localities.R
\name{road_get_localities}
\alias{road_get_localities}
\title{Get localities from ROAD Database}
\usage{
road_get_localities(
  continents = NULL,
  subcontinents = NULL,
  countries = NULL,
  locality_types = NULL,
  cultural_periods = NULL,
  technocomplexes = NULL
)
}
\arguments{
\item{continents}{specifies the continent(s) (e.g. Africa, Europe, Asia).
Run \code{road_list_argument_values("continents")} to display possible values.
The argument\code{continents} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{subcontinents}{specifies the continental region(s) (e.g. Southern Europe). 
Run \code{road_list_argument_values("subcontinents")} to display possible values.
The argument \code{subcontinents} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{countries}{specifies the name of the country where a locality is situated 
(e.g. Germany, Kenya, Saudi Arabia). Run \code{road_list_argument_values("countries")} 
to display possible values.
The argument \code{countries} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{locality_types}{specifies the type of locality (e.g. cave, rockshelter, open air).
Run \code{road_list_argument_values("locality_types")} to display possible values.
The argument \code{locality_types} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{cultural_periods}{specifies the main cultural epoch(s) and includes the 
Eurasian Paleolithic (Lower, Middle, Upper, Epi-) and the African Stone Age 
(Earlier, Middle, Later). Run \code{road_list_argument_values("cultural_periods")} 
to display possible values. The argument \code{cultural_periods} is a string 
(one item) or vector of strings (one or more items); defaults to NULL.}

\item{technocomplex}{specifies an archaeological culture or named stone tool 
industry (e.g. Oldowan, Acheulean, Mousterian).
Run \code{road_list_argument_values("technocomplex")} to display possible values.
The argument \code{technocomplex} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}
}
\value{
Database search result as a data frame with the information about localities 
like their geographic information, cultural period and type.
}
\description{
The \strong{\code{road_get_localities}} fetches data of archaeological sites 
(localities) from ROAD database. The ROAD table locality provides basic 
information about each place where an assemblage of archaeological, 
paleoanthropological, paleontological, paleobotanical or other relevant materials
was described, recorded, sampled or collected. Every locality (site) is situated in a specific
country within a given geographic region. The name of every locality is unique.
}
\details{
Use arguments to filter search results by location, type, or culture or omit 
them to have a broader result set. All arguments are optional and should be 
omitted or set to NULL when not used.
}
\examples{
df <- road_get_localities(continents = "Europe", countries = c("Germany"), 
                    locality_type = c("basin", "quarry"))
df <- road_get_localities(countries = c("Germany", "France"), cultural_periods = "Epipaleolithic")
}
