% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_localities.R
\name{road_get_localities_internal}
\alias{road_get_localities_internal}
\title{df <- road_get_localities_internal(countries =
                c("Germany", "France"), cultural_periods = "Epipaleolithic")}
\usage{
road_get_localities_internal(
  continents = NULL,
  subcontinents = NULL,
  countries = NULL,
  locality_types = NULL,
  cultural_periods = NULL,
  technocomplexes = NULL
)
}
\description{
df <- road_get_localities_internal(countries =
                c("Germany", "France"), cultural_periods = "Epipaleolithic")
}
