% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_archaeology.R
\name{road_get_miscellaneous_finds}
\alias{road_get_miscellaneous_finds}
\title{Get miscellaneous finds from ROAD database}
\usage{
road_get_miscellaneous_finds(
  continents = NULL,
  subcontinents = NULL,
  countries = NULL,
  locality_types = NULL,
  cultural_periods = NULL,
  technocomplexes = NULL,
  categories = NULL,
  age_min = NULL,
  age_max = NULL,
  miscellaneous_finds_material = NULL,
  assemblages = NULL
)
}
\arguments{
\item{continents}{specifies the continent(s) (e.g. Africa, Europe, Asia).
Run \code{road_list_argument_values("continents")} to display possible values.
The argument\code{continents} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{subcontinents}{specifies the continental region(s) (e.g. Southern Europe). 
Run \code{road_list_argument_values("subcontinents")} to display possible values.
The argument \code{subcontinents} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{countries}{specifies the name of the country where a locality is situated 
(e.g. Germany, Kenya, Saudi Arabia). Run \code{road_list_argument_values("countries")} 
to display possible values.
The argument \code{countries} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{locality_types}{specifies the type of locality (e.g. cave, rockshelter, open air).
Run \code{road_list_argument_values("locality_types")} to display possible values.
The argument \code{locality_types} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{cultural_periods}{specifies the main cultural epoch(s) and includes the 
Eurasian Paleolithic (Lower, Middle, Upper, Epi-) and the African Stone Age 
(Earlier, Middle, Later). Run \code{road_list_argument_values("cultural_periods")} 
to display possible values. The argument \code{cultural_periods} is a string 
(one item) or vector of strings (one or more items); defaults to NULL.}

\item{technocomplexes}{specifies an archaeological culture or named stone tool 
industry (e.g. Oldowan, Acheulean, Mousterian).
Run \code{road_list_argument_values("technocomplexes")} to display possible values.
The argument \code{technocomplexes} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{categories}{specifies the assemblage category/categories with the classes 
human remains, raw material, typology, technology, function, organic tools, 
symbolic artifacts, feature, miscellaneous finds, paleofauna, animal remains, 
plant remains. The argument \code{categories} is a string (one item) or 
vector of strings (one or more items); defaults to NULL.}

\item{age_min}{specifies the minimum age in years before present, using 1950 CE 
as the baseline. The argument \code{age_min} is an integer; defaults to NULL.}

\item{age_max}{specifies the maximum age in years before present, using 1950 CE 
as the baseline. The argument \code{age_max} is an integer; defaults to NULL.}

\item{miscellaneous_finds_material}{specifies material of the miscellaneous 
finds (e.g. shell, ochre, ostrich eggshell).
Run \code{road_list_argument_values("miscellaneous_finds_material")} to 
display possible values. The argument \code{miscellaneous_finds_material} is 
a string (one item) or vector of strings; defaults to NULL.}

\item{assemblages}{list of assemblages; return value from function `road_get_assemblages`.
Can be used instead of the other locality and assemblage parameters to filter the results.}
}
\value{
Database search result as a data frame with the information about 
miscellaneous finds like their geographic information,
cultural period, locality type, category, dating, miscellaneous finds material,
miscellaneous finds material source (local, regional, supra-regional, unknown) and number.
}
\description{
The \strong{\code{road_get_miscellaneous_finds}} fetches data of feature finds from ROAD database.
Miscellaneous finds are archaeological objects that do not fit into other specific categories.
Miscellaneous finds are classified by their material.
This function allows you to query miscellaneous finds data from the ROAD database
using parameters such as geographical location,
cultural periods, material types, and assemblages. Use the parameters to filter the results
or omit them to retrieve a broader dataset.
}
\examples{
road_get_miscellaneous_finds(miscellaneous_finds_material = "wood fossil")
road_get_miscellaneous_finds(continents = c("Africa"), locality_types = "open air",
                             miscellaneous_finds_material = "shell")
}
