% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_archaeology.R
\name{road_get_miscellaneous_finds}
\alias{road_get_miscellaneous_finds}
\title{Get miscellaneous finds from ROAD database}
\usage{
road_get_miscellaneous_finds(
  continents = NULL,
  subcontinents = NULL,
  countries = NULL,
  locality_types = NULL,
  cultural_periods = NULL,
  categories = NULL,
  age_min = NULL,
  age_max = NULL,
  miscellaneous_finds_material = NULL,
  assemblages = NULL
)
}
\arguments{
\item{continents}{specifies the continent(s) of the country/countries, e.g. Africa, Europe, Asia. The parameter continents is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{subcontinents}{specifies the continental region(s) of the country , e.g. Southern Europe. The parameter subcontinents is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{countries}{specifies the name of the country where a locality is situated (e.g.
Germany, Kenya, Saudi Arabia, China). The parameter countries is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{locality_types}{specifies the type of locality (e.g. cave, rockshelter, open air, profile, outcrop,
mine, quarry, boring). The parameter locality_types is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{cultural_periods}{specifies the main cultural epoch(s) and includes the Eurasian
Paleolithic (Lower, Middle, Upper, Epi) and the African Stone Age (Earlier, Middle, Later). The parameter cultural_periods is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{categories}{specifies the assemblage category/categories. For an assemblage consisting of human remains category includes the entry “human
remains”. In the case of archaeological assemblages, multiple categories are the norm and may
include “raw material, typology, technology, function, organic tools, symbolic artifacts, feature,
miscellaneous finds”. A faunal assemblage can also contain multiple entries including
“paleofauna, animal remains”, while a botanical assemblage can only include the entry “plant
remains”.The parameter categories is a string (one item) or vector of strings (one or more items).}

\item{age_min}{specifies the minimum age of assemblage. The parameter age_min is an integer.}

\item{age_max}{specifies the maximum age of assemblage. The parameter age_max is an integer.}

\item{miscellaneous_finds_material}{specifies material of the miscellaneous find 
and may include: beeswax, bitumen, clay, flax fiber, fossil, mineral diverse 
(materials such as crystals), mineral pigment (colorants such as ochre, hematite, 
limonite, goethite, specularite, etc.) ostrich eggshell, shell,
stone, etc.  The parameter miscellaneous_finds_material is 
a string (one item) or vector of strings}

\item{assemblages}{list of assemblages; return value from function `road_get_assemblages`.}
}
\value{
Database search result as list of miscellaneous finds with their geographic information, 
cultural period, locality type, category, dating, miscellaneous finds material, 
miscellaneous finds material source (local, regional, supra-regional, unknown) and number.
}
\description{
`road_get_miscellaneous_finds` fetches data of feature finds from ROAD database.
Miscellaneous finds are archaeological objects that do not fit into other specific categories. 
Miscellaneous finds are classified by their material.
This function allows you to query miscellaneous finds data from the ROAD database 
using parameters such as geographical location,
cultural periods, material types, and assemblages. Use the parameters to filter the results 
or omit them to retrieve a broader dataset.
}
\examples{
road_get_miscellaneous_finds(miscellaneous_finds_material = "wood fossil")
road_get_miscellaneous_finds(continents = c("Africa"), locality_types = "open air",  
                             miscellaneous_finds_material = "shell")
}
