% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/road_package_aux.R
\name{road_get_miscellaneous_finds}
\alias{road_get_miscellaneous_finds}
\title{Get miscellaneous finds from ROAD database}
\usage{
road_get_miscellaneous_finds(
  continents = NULL,
  subcontinents = NULL,
  countries = NULL,
  locality_types = NULL,
  cultural_periods = NULL,
  categories = NULL,
  age_min = NULL,
  age_max = NULL,
  miscellaneous_finds_material = NULL,
  assemblages = NULL
)
}
\arguments{
\item{continents}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{subcontinents}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{countries}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{locality_types}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{cultural_periods}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{categories}{string (one item) or vector of strings (one or more items).}

\item{age_min}{integer; minimum age of assemblage.}

\item{age_max}{integer; maximum age of assemblage.}

\item{miscellaneous_finds_material}{string (one item) or vector of strings}

\item{assemblages}{list of assemblages; return value from function `road_get_assemblages`.}
}
\value{
Database search result as list of miscellaneous finds.
}
\description{
`road_get_miscellaneous_finds` fetches data of feature finds from ROAD database.
}
\examples{
road_get_miscellaneous_finds(continents = "Europe")
road_get_miscellaneous_finds(continents = "Europe", )
road_get_miscellaneous_finds(continents = "Europe", )
}
