% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_archaeology.R
\name{road_get_organic_tools}
\alias{road_get_organic_tools}
\title{Get organic tools from ROAD database}
\usage{
road_get_organic_tools(
  continents = NULL,
  subcontinents = NULL,
  countries = NULL,
  locality_types = NULL,
  cultural_periods = NULL,
  technocomplexes = NULL,
  categories = NULL,
  age_min = NULL,
  age_max = NULL,
  organic_tools_interpretation = NULL,
  assemblages = NULL
)
}
\arguments{
\item{continents}{specifies the continent(s) (e.g. Africa, Europe, Asia).
Run \code{road_list_argument_values("continents")} to display possible values.
The argument\code{continents} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{subcontinents}{specifies the continental region(s) (e.g. Southern Europe). 
Run \code{road_list_argument_values("subcontinents")} to display possible values.
The argument \code{subcontinents} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{countries}{specifies the name of the country where a locality is situated 
(e.g. Germany, Kenya, Saudi Arabia). Run \code{road_list_argument_values("countries")} 
to display possible values.
The argument \code{countries} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{locality_types}{specifies the type of locality (e.g. cave, rockshelter, open air).
Run \code{road_list_argument_values("locality_types")} to display possible values.
The argument \code{locality_types} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{cultural_periods}{specifies the main cultural epoch(s) and includes the 
Eurasian Paleolithic (Lower, Middle, Upper, Epi-) and the African Stone Age 
(Earlier, Middle, Later). Run \code{road_list_argument_values("cultural_periods")} 
to display possible values. The argument \code{cultural_periods} is a string 
(one item) or vector of strings (one or more items); defaults to NULL.}

\item{technocomplexes}{specifies an archaeological culture or named stone tool 
industry (e.g. Oldowan, Acheulean, Mousterian).
Run \code{road_list_argument_values("technocomplexes")} to display possible values.
The argument \code{technocomplexes} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{categories}{specifies the assemblage category/categories with the classes 
human remains, raw material, typology, technology, function, organic tools, 
symbolic artifacts, feature, miscellaneous finds, paleofauna, animal remains, 
plant remains. The argument \code{categories} is a string (one item) or 
vector of strings (one or more items); defaults to NULL.}

\item{age_min}{specifies the minimum age in years before present, using 1950 CE 
as the baseline. The argument \code{age_min} is an integer; defaults to NULL.}

\item{age_max}{specifies the maximum age in years before present, using 1950 CE 
as the baseline. The argument \code{age_max} is an integer; defaults to NULL.}

\item{organic_tools_interpretation}{specifies interpreted organic tool types 
(e.g. lance/spear, point, retoucher). 
Run \code{road_list_argument_values("organic_tools_interpretation")} 
to display possible values. The argument \code{organic_tools_interpretation} 
is a string (one item) or vector of strings; defaults to NULL.}

\item{assemblages}{list of assemblages; return value from function `road_get_assemblages`.
Can be used instead of the other locality and assemblage parameters to filter the results.}
}
\value{
Database search result as a data frame with the information about 
organic tools like their geographic information,
cultural period, locality type, category, dating, information about
organic tools interpretation, organic raw material, organic tools technology and number.
}
\description{
The \strong{\code{road_get_organic_tools}} fetches data of organic tools from 
ROAD database.Organic tools are artifacts made from organic materials such as 
bone, antler, or wood, found in archaeological contexts. This function enables 
you to query organic tool data from the ROAD database based on parameters like 
geographical location, cultural periods, tool interpretation, and assemblages. 
Use the parameters to filter the results or omit them for broader results.
}
\examples{
road_get_organic_tools(continents = c("Europe"), organic_tools_interpretation = "fishhook")
road_get_organic_tools(continents = "Africa", organic_tools_interpretation = c("fishhook"))
}
