% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_archaeology.R
\name{road_get_organic_tools}
\alias{road_get_organic_tools}
\title{Get organic tools from ROAD database}
\usage{
road_get_organic_tools(
  continents = NULL,
  subcontinents = NULL,
  countries = NULL,
  locality_types = NULL,
  cultural_periods = NULL,
  categories = NULL,
  age_min = NULL,
  age_max = NULL,
  organic_tools_interpretation = NULL,
  assemblages = NULL
)
}
\arguments{
\item{continents}{specifies the continent(s) of the country/countries, e.g. Africa, Europe, Asia. The parameter continents is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{subcontinents}{specifies the continental region(s) of the country , e.g. Southern Europe. The parameter subcontinents is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{countries}{specifies the name of the country where a locality is situated (e.g.
Germany, Kenya, Saudi Arabia, China). The parameter countries is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{locality_types}{specifies the type of locality (e.g. cave, rockshelter, open air, profile, outcrop,
mine, quarry, boring). The parameter locality_types is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{cultural_periods}{specifies the main cultural epoch(s) and includes the Eurasian
Paleolithic (Lower, Middle, Upper, Epi) and the African Stone Age (Earlier, Middle, Later). The parameter cultural_periods is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{categories}{specifies the assemblage category/categories. For an assemblage consisting of human remains category includes the entry “human
remains”. In the case of archaeological assemblages, multiple categories are the norm and may
include “raw material, typology, technology, function, organic tools, symbolic artifacts, feature,
miscellaneous finds”. A faunal assemblage can also contain multiple entries including
“paleofauna, animal remains”, while a botanical assemblage can only include the entry “plant
remains”.The parameter categories is a string (one item) or vector of strings (one or more items).}

\item{age_min}{specifies the minimum age of assemblage. The parameter age_min is an integer.}

\item{age_max}{specifies the maximum age of assemblage. The parameter age_max is an integer.}

\item{organic_tools_interpretation}{specifies interpreted organic tool types. Possible values for the
interpretation are: lance/spear, needle eyed, percussor, pin, point, point barbed, 
point split based, pointed artifact, punch, retoucher, rod, scraper, 
shaftstring (one item) or vector of strings}

\item{assemblages}{list of assemblages; return value from function `road_get_assemblages`.}
}
\value{
Database search result as list of organic tools with their geographic information, 
cultural period, locality type, category, dating, information about 
organic tools interpretation, organic raw material, organic tools technology and number.
}
\description{
`road_get_organic_tools` fetches data of organic tools from ROAD database.Organic 
tools are artifacts made from organic materials such as bone, antler, or wood, 
found in archaeological contexts. This function enables you to query organic tool 
data from the ROAD database based on parameters like geographical location,
cultural periods, tool interpretation, and assemblages. Use the parameters 
to filter the results or omit them for broader results.
}
\examples{
road_get_organic_tools(continents = c("Europe"), organic_tools_interpretation = "fishhook")
road_get_organic_tools(continents = "Africa", organic_tools_interpretation = c("fishhook"))
}
