% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paleofauna.R
\name{road_get_paleofauna}
\alias{road_get_paleofauna}
\title{Get paleofauna finds from ROAD database}
\usage{
road_get_paleofauna(
  continents = NULL,
  subcontinents = NULL,
  countries = NULL,
  locality_types = NULL,
  cultural_periods = NULL,
  categories = NULL,
  age_min = NULL,
  age_max = NULL,
  assemblages = NULL,
  fauna_genus = NULL,
  fauna_species = NULL
)
}
\arguments{
\item{continents}{specifies the continent(s) of the country/countries, e.g. Africa, Europe, Asia. The parameter continents is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{subcontinents}{specifies the continental region(s) of the country , e.g. Southern Europe. The parameter subcontinents is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{countries}{specifies the name of the country where a locality is situated (e.g.
Germany, Kenya, Saudi Arabia, China). The parameter countries is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{locality_types}{specifies the type of locality (e.g. cave, rockshelter, open air, profile, outcrop,
mine, quarry, boring). The parameter locality_types is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{cultural_periods}{specifies the main cultural epoch(s) and includes the Eurasian
Paleolithic (Lower, Middle, Upper, Epi) and the African Stone Age (Earlier, Middle, Later). The parameter cultural_periods is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{categories}{specifies the assemblage category/categories. For an assemblage consisting of human remains category includes the entry “human
remains”. In the case of archaeological assemblages, multiple categories are the norm and may
include “raw material, typology, technology, function, organic tools, symbolic artifacts, feature,
miscellaneous finds”. A faunal assemblage can also contain multiple entries including
“paleofauna, animal remains”, while a botanical assemblage can only include the entry “plant
remains”.The parameter categories is a string (one item) or vector of strings (one or more items).}

\item{age_min}{specifies the minimum age of assemblage. The parameter age_min is an integer.}

\item{age_max}{specifies the maximum age of assemblage. The parameter age_max is an integer.}

\item{assemblages}{list of assemblages; return value from function `road_get_assemblages`.
Can be used instead of the other locality and assemblage parameters to filter the results.}

\item{fauna_genus}{specifies the genus to which the described faunal remains is attributed to.
Possible entries include: "Mammuthus", "Vulpes" etc.
The parameter genus is a string (one item) or vector of strings (one or more items).}

\item{fauna_species}{specifies the species to which the described faunal remains is attributed. Possible entries include:
"primigenius", "vulpes" or "sp." for unidentified species.
The parameter species is a string (one item) or vector of strings (one or more items).}
}
\value{
Database search result as list of assemblages with paleofauna finds.
}
\description{
`road_get_paleofauna` fetches data of paleofauna finds from the ROAD database.
}
\details{
Paleofauna finds are animal fossil remains discovered in archaeological contexts and are always associated with an assemblage.
These finds provide direct evidence for the presence of animal species at a particular locality and time.
The function returns information about the assemblage in which certain faunal remains were found as well as their genus and species.
Use the parameters to filter the results or omit them to retrieve a broader dataset.
Genus and species parameters can be entered as a vector of strings to search for multiple entries.
If genus and species are both specified, most of the time it's more sensible to enter them as
single strings and not as vectors with multiple search words to recieve useful results.
}
\examples{
road_get_paleofauna(fauna_genus = "Mammuthus")
road_get_paleofauna(continents = "Europe", fauna_genus = "Vulpes", fauna_species = "vulpes")
}
