% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paleofauna.R
\name{road_get_paleofauna}
\alias{road_get_paleofauna}
\title{Get paleofauna finds from ROAD database}
\usage{
road_get_paleofauna(
  assemblages = NULL,
  continents = NULL,
  subcontinents = NULL,
  countries = NULL,
  locality_types = NULL,
  cultural_periods = NULL,
  categories = NULL,
  age_min = NULL,
  age_max = NULL,
  fauna_genus = NULL,
  fauna_species = NULL
)
}
\arguments{
\item{assemblages}{list of assemblages; return value from function `road_get_assemblages`.}

\item{continents}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{subcontinents}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{countries}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{locality_types}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{cultural_periods}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{categories}{string (one item) or vector of strings (one or more items).}

\item{age_min}{integer; minimum age of assemblage.}

\item{age_max}{integer; maximum age of assemblage.}

\item{fauna_genus}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{fauna_species}{string (one item) or vector of strings (one or more items); defaults to NULL.}
}
\value{
Database search result as list of archaeological finds.
}
\description{
`road_get_paleofauna` fetches data of paleofauna finds from ROAD database.
}
\details{
Paleofauna finds are often part of an assemblage which means the function needs a list of
assemblages (return value of function `road_get_assemblages`) as its parameter.
}
\examples{
road_get_paleofauna(continents = "Europe")
road_get_paleofauna(continents = "Europe", fauna_genus = "Vulpes", fauna_species = "vulpes")
}
