% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paleofauna.R
\name{road_get_paleofauna}
\alias{road_get_paleofauna}
\title{Get paleofauna finds from ROAD database}
\usage{
road_get_paleofauna(
  continents = NULL,
  subcontinents = NULL,
  countries = NULL,
  locality_types = NULL,
  cultural_periods = NULL,
  categories = NULL,
  age_min = NULL,
  age_max = NULL,
  assemblages = NULL,
  fauna_genus = NULL,
  fauna_species = NULL
)
}
\arguments{
\item{continents}{specifies the continent(s) (e.g. Africa, Europe, Asia).
Run \code{road_list_argument_values("continents")} to display possible values.
The argument\code{continents} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{subcontinents}{specifies the continental region(s) (e.g. Southern Europe). 
Run \code{road_list_argument_values("subcontinents")} to display possible values.
The argument \code{subcontinents} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{countries}{specifies the name of the country where a locality is situated 
(e.g. Germany, Kenya, Saudi Arabia). Run \code{road_list_argument_values("countries")} 
to display possible values.
The argument \code{countries} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{locality_types}{specifies the type of locality (e.g. cave, rockshelter, open air).
Run \code{road_list_argument_values("locality_types")} to display possible values.
The argument \code{locality_types} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{cultural_periods}{specifies the main cultural epoch(s) and includes the 
Eurasian Paleolithic (Lower, Middle, Upper, Epi-) and the African Stone Age 
(Earlier, Middle, Later). Run \code{road_list_argument_values("cultural_periods")} 
to display possible values. The argument \code{cultural_periods} is a string 
(one item) or vector of strings (one or more items); defaults to NULL.}

\item{categories}{specifies the assemblage category/categories with the classes 
human remains, raw material, typology, technology, function, organic tools, 
symbolic artifacts, feature, miscellaneous finds, paleofauna, animal remains, 
plant remains. The argument \code{categories} is a string (one item) or 
vector of strings (one or more items); defaults to NULL.}

\item{age_min}{specifies the minimum age in years before present, using 1950 CE 
as the baseline. The argument \code{age_min} is an integer; defaults to NULL.}

\item{age_max}{specifies the maximum age in years before present, using 1950 CE 
as the baseline. The argument \code{age_max} is an integer; defaults to NULL.}

\item{assemblages}{list of assemblages; return value from function \code{road_get_assemblages}.
Can be used instead of the other locality and assemblage parameters to filter the results.}

\item{fauna_genus}{specifies the genus to which the described faunal remains 
is attributed to. Possible entries include: "Mammuthus", "Vulpes" etc.
Run \code{road_list_argument_values("fauna_genus")} to
display possible values. The argument \code{fauna_genus} is a string 
(one item) or vector of strings; defaults to NULL.}

\item{fauna_species}{specifies the species to which the 
described faunal remains is attributed. Possible entries include:
"primigenius", "vulpes" or "sp." for unidentified species. Run 
\code{road_list_argument_values("fauna_species")} to display possible values.
The argument \code{fauna_species} is a string (one item) or vector of strings; 
defaults to NULL.}

\item{technocomplex}{specifies an archaeological culture or named stone tool 
industry (e.g. Oldowan, Acheulean, Mousterian).
Run \code{road_list_argument_values("technocomplex")} to display possible values.
The argument \code{technocomplex} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}
}
\value{
Database search result as list of assemblages with paleofauna information.
}
\description{
The  \strong{\code{road_get_paleofauna}} function fetches data of paleofauna finds from the ROAD database.
Paleofauna finds are animal fossil remains discovered in archaeological contexts and are always associated with an assemblage.
These finds provide direct evidence for the presence of animal species at a particular locality and time.
The function returns information about the assemblage in which certain faunal remains were found as well as their genus and species.
}
\details{
Use the parameters to filter the results or omit them to retrieve a broader dataset.
Genus and species parameters can be entered as a vector of strings to search for multiple entries.
If genus and species are both specified, most of the time it's more sensible to enter them as
single strings and not as vectors with multiple search words to recieve useful results.
}
\examples{
road_get_paleofauna(fauna_genus = "Mammuthus")
road_get_paleofauna(continents = "Europe", fauna_genus = "Vulpes", 
                    fauna_species = "vulpes")
}
