% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plantremains.R
\name{road_get_plantremains}
\alias{road_get_plantremains}
\title{Get paleobotany data from ROAD database}
\usage{
road_get_plantremains(
  assemblages = NULL,
  continents = NULL,
  subcontinents = NULL,
  countries = NULL,
  locality_types = NULL,
  cultural_periods = NULL,
  categories = NULL,
  age_min = NULL,
  age_max = NULL,
  plant_remains = NULL,
  plant_family = NULL,
  plant_genus = NULL,
  plant_species = NULL
)
}
\arguments{
\item{assemblages}{list of assemblages; return value from function `road_get_assemblages`.}

\item{continents}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{subcontinents}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{countries}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{locality_types}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{cultural_periods}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{categories}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{age_min}{integer; minimum age of paleobotanical remains.}

\item{age_max}{integer; maximum age of paleobotanical remains.}

\item{plant_remains}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{plant_family}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{plant_genus}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{plant_species}{string (one item) or vector of strings (one or more items); defaults to NULL.}
}
\value{
Database search result as a list of assemblages with paleobotanical remains.
}
\description{
`road_get_plantremains` fetches data of paleobotanical remains from the ROAD database.
}
\details{
Paleobotanical remains are plant remains found in archaeological contexts. This function allows you to query
paleobotanical data based on various parameters such as geographical location, cultural periods, plant taxonomy,
and assemblages. Use the parameters to filter the results or omit them to retrieve broader results.
}
\examples{
road_get_plantremains(plant_genus = "Triticum")
road_get_plantremains(categories = "plant remains", age_min = 5000L, age_max = 10000L)
road_get_plantremains(countries = c("Germany", "France"), plant_family = "Poaceae")
}
