% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_archaeology.R
\name{road_get_symbolic_artifacts}
\alias{road_get_symbolic_artifacts}
\title{Get symbolic artifacts from ROAD database}
\usage{
road_get_symbolic_artifacts(
  continents = NULL,
  subcontinents = NULL,
  countries = NULL,
  locality_types = NULL,
  cultural_periods = NULL,
  technocomplexes = NULL,
  categories = NULL,
  age_min = NULL,
  age_max = NULL,
  symbolic_artifacts_interpretation = NULL,
  assemblages = NULL
)
}
\arguments{
\item{continents}{specifies the continent(s) (e.g. Africa, Europe, Asia).
Run \code{road_list_argument_values("continents")} to display possible values.
The argument \code{continents} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{subcontinents}{specifies the continental region(s) (e.g. Southern Europe). 
Run \code{road_list_argument_values("subcontinents")} to display possible values.
The argument \code{subcontinents} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{countries}{specifies the name of the country where a locality is situated 
(e.g. Germany, Kenya, Saudi Arabia). Run \code{road_list_argument_values("countries")} 
to display possible values.
The argument \code{countries} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{locality_types}{specifies the type of locality (e.g. cave, rockshelter, open air).
Run \code{road_list_argument_values("locality_types")} to display possible values.
The argument \code{locality_types} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{cultural_periods}{specifies the main cultural epoch(s) and includes the 
Eurasian Paleolithic (Lower, Middle, Upper, Epi-) and the African Stone Age 
(Earlier, Middle, Later). Run \code{road_list_argument_values("cultural_periods")} 
to display possible values. The argument \code{cultural_periods} is a string 
(one item) or vector of strings (one or more items); defaults to NULL.}

\item{technocomplexes}{specifies an archaeological culture or named stone tool 
industry (e.g. Oldowan, Acheulean, Mousterian).
Run \code{road_list_argument_values("technocomplexes")} to display possible values.
The argument \code{technocomplexes} is a string (one item) or vector of strings 
(one or more items); defaults to NULL.}

\item{categories}{specifies the assemblage category/categories with the classes 
human remains, raw material, typology, technology, function, organic tools, 
symbolic artifacts, feature, miscellaneous finds, paleofauna, animal remains, 
plant remains. The argument \code{categories} is a string (one item) or 
vector of strings (one or more items); defaults to NULL.}

\item{age_min}{specifies the minimum age in years before present, using 1950 CE 
as the baseline. The argument \code{age_min} is an integer; defaults to NULL.}

\item{age_max}{specifies the maximum age in years before present, using 1950 CE 
as the baseline. The argument \code{age_max} is an integer; defaults to NULL.}

\item{symbolic_artifacts_interpretation}{specifies the interpretation of 
symbolic artifacts (e.g. abstract, anthropomorphic, zoomorphic, instrument, 
ornament). Run \code{road_list_argument_values("symbolic_artifacts_interpretation")} 
to display possible values. The argument \code{symbolic_artifacts_interpretation} 
is a string (one item) or vector of strings; defaults to NULL.}

\item{assemblages}{list of assemblages; return value from function 
\code{road_get_assemblages}. Can be used instead of the other locality and 
assemblage parameters to filter the results.}
}
\value{
Database search result as a data frame with the information about 
symbolic artifacts like their geographic information, cultural period, 
locality type, assemblage category, dating and info about symbolic artifacts 
interpretation, symbolic artifacts category, symbolic artifacts technology, 
symbolic artifacts material, symbolic artifacts raw material source.
}
\description{
The \strong{\code{road_get_symbolic_artifacts}} fetches data of symbolic artifacts from ROAD database.
Symbolic artifacts are objects interpreted as having symbolic or cultural significance
in archaeological contexts. This function allows you to query symbolic artifact
data from the ROAD database using parameters such as geographical location,
cultural periods, artifact interpretation, and assemblages. Use the parameters
to filter the results or omit them to retrieve a broader dataset.
}
\examples{
road_get_symbolic_artifacts(symbolic_artifacts_interpretation = c("instrument"))
road_get_symbolic_artifacts(continents = "Africa", symbolic_artifacts_interpretation = "zoomorphic")
}
