% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_archaeology.R
\name{road_get_symbolic_artifacts}
\alias{road_get_symbolic_artifacts}
\title{Get symbolic artifacts from ROAD database}
\usage{
road_get_symbolic_artifacts(
  continents = NULL,
  subcontinents = NULL,
  countries = NULL,
  locality_types = NULL,
  cultural_periods = NULL,
  categories = NULL,
  age_min = NULL,
  age_max = NULL,
  symbolic_artifacts_interpretation = NULL,
  assemblages = NULL
)
}
\arguments{
\item{continents}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{subcontinents}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{countries}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{locality_types}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{cultural_periods}{string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{categories}{string (one item) or vector of strings (one or more items).}

\item{age_min}{integer; minimum age of assemblage.}

\item{age_max}{integer; maximum age of assemblage.}

\item{symbolic_artifacts_interpretation}{string (one item) or vector of strings}

\item{assemblages}{list of assemblages; return value from function `road_get_assemblages`.}
}
\value{
Database search result as list of symbolic artifacts with info about symbolic.
}
\description{
`road_get_symbolic_artifacts` fetches data of symbolic artifacts from ROAD database.
}
\examples{
road_get_symbolic_artifacts(continents = "Europe")
road_get_symbolic_artifacts(continents = "Europe", )
road_get_symbolic_artifacts(continents = "Europe", )
}
