% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_archaeology.R
\name{road_get_symbolic_artifacts}
\alias{road_get_symbolic_artifacts}
\title{Get symbolic artifacts from ROAD database}
\usage{
road_get_symbolic_artifacts(
  continents = NULL,
  subcontinents = NULL,
  countries = NULL,
  locality_types = NULL,
  cultural_periods = NULL,
  categories = NULL,
  age_min = NULL,
  age_max = NULL,
  symbolic_artifacts_interpretation = NULL,
  assemblages = NULL
)
}
\arguments{
\item{continents}{specifies the continent(s) of the country/countries, e.g. Africa, Europe, Asia. The parameter continents is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{subcontinents}{specifies the continental region(s) of the country , e.g. Southern Europe. The parameter subcontinents is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{countries}{specifies the name of the country where a locality is situated (e.g.
Germany, Kenya, Saudi Arabia, China). The parameter countries is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{locality_types}{specifies the type of locality (e.g. cave, rockshelter, open air, profile, outcrop,
mine, quarry, boring). The parameter locality_types is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{cultural_periods}{specifies the main cultural epoch(s) and includes the Eurasian
Paleolithic (Lower, Middle, Upper, Epi) and the African Stone Age (Earlier, Middle, Later). The parameter cultural_periods is a string (one item) or vector of strings (one or more items); defaults to NULL.}

\item{categories}{specifies the assemblage category/categories. For an assemblage consisting of human remains category includes the entry “human
remains”. In the case of archaeological assemblages, multiple categories are the norm and may
include “raw material, typology, technology, function, organic tools, symbolic artifacts, feature,
miscellaneous finds”. A faunal assemblage can also contain multiple entries including
“paleofauna, animal remains”, while a botanical assemblage can only include the entry “plant
remains”.The parameter categories is a string (one item) or vector of strings (one or more items).}

\item{age_min}{specifies the minimum age of assemblage. The parameter age_min is an integer.}

\item{age_max}{specifies the maximum age of assemblage. The parameter age_max is an integer.}

\item{symbolic_artifacts_interpretation}{specifies symbolic artifacts. Possible values for the
interpretation can be abstract, anthropomorphic, instrument, ornament, therianthropic,
zoomorphic, etc. The parameter symbolic_artifacts_interpretation is a string (one item)
or vector of strings}

\item{assemblages}{list of assemblages; return value from function `road_get_assemblages`.
Can be used instead of the other locality and assemblage parameters to filter the results.}
}
\value{
Database search result as a data frame with the information about symbolic artifacts 
like their geographic information, cultural period, locality type, assemblage category, 
dating and info about symbolic artifacts interpretation, symbolic artifacts 
category, symbolic artifacts technology, symbolic artifacts material, symbolic artifacts raw material source.
}
\description{
The \strong{\code{road_get_symbolic_artifacts}} fetches data of symbolic artifacts from ROAD database.
Symbolic artifacts are objects interpreted as having symbolic or cultural significance
in archaeological contexts. This function allows you to query symbolic artifact
data from the ROAD database using parameters such as geographical location,
cultural periods, artifact interpretation, and assemblages. Use the parameters
to filter the results or omit them to retrieve a broader dataset.
}
\examples{
road_get_symbolic_artifacts(symbolic_artifacts_interpretation = c("instrument"))
road_get_symbolic_artifacts(continents = "Africa", symbolic_artifacts_interpretation = "zoomorphic")
}
