% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{road_list_argument_values}
\alias{road_list_argument_values}
\title{Get argument value from ROAD Database}
\usage{
road_list_argument_values(function_parameter, function_name = NULL)
}
\arguments{
\item{function_parameter}{name of a function parameter}

\item{function_name}{name of a function}
}
\value{
List of attribute values.
}
\description{
The  \strong{\code{road_list_argument_values}} function fetches values of a 
given parameter in the database or read values from file
}
\examples{
road_list_argument_values("locality_types")
}
