% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{road_list_argument_values}
\alias{road_list_argument_values}
\title{Get argument value from ROAD Database}
\usage{
road_list_argument_values(function_argument)
}
\arguments{
\item{function_argument}{name of a function argument}
}
\value{
List of attribute values.
}
\description{
The  \strong{\code{road_list_argument_values}} function fetches values of a 
given argument in the database or read values from file
}
\examples{
road_list_argument_values("locality_type")
road_list_argument_values("subcontinent")
road_list_argument_values("dating_methods")
}
