% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{road_list_values}
\alias{road_list_values}
\title{Get attribute value from ROAD Database}
\usage{
road_list_values(attribute_name = NULL)
}
\arguments{
\item{attribute_name}{name of an attribute; defaults to NULL.}
}
\value{
List of attribute values.
}
\description{
`road_list_values` fetches values of a given attribute in the database or 
read values from file
}
\details{
#' All parameters are optional and should be omitted or set to NULL when not used.
}
\examples{
road_list_values("category")
road_list_values("cultural_period")
}
