% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/road_package_aux.R
\name{road_list_values}
\alias{road_list_values}
\title{Get attribute value from ROAD Database}
\usage{
road_list_values(attribute_name = NULL)
}
\arguments{
\item{continents}{string (one item) or vector of strings (one or more items); defaults to NULL.}
}
\value{
List of attribute values.
}
\description{
`road_list_values` fetches values of a given attribute
}
\examples{
road_list_values("category")
road_list_values("cultural_period")
}
