% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_funcs.R
\name{road_run_query}
\alias{road_run_query}
\title{run query in ROAD database}
\usage{
road_run_query(query)
}
\arguments{
\item{query}{specifies the SQl query.}
}
\value{
Database search result as a data frame.
}
\description{
run query in ROAD database
}
\keyword{'transport_distance',}
\keyword{a}
\keyword{correct}
\keyword{dont}
\keyword{internal}
\keyword{is}
\keyword{it.}
\keyword{of}
\keyword{replace}
\keyword{unknown'}
\keyword{value}
\keyword{want}
\keyword{we}
