% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{road_summarize_archaeology}
\alias{road_summarize_archaeology}
\title{Get an Overview of the ROAD Archaeology Database}
\usage{
road_summarize_archaeology(term)
}
\arguments{
\item{term}{A string containing the search term (one item).}
}
\value{
Database search result as list of lithic finds with info about typology.
}
\description{
The `road_summarize_archaeology` function provides a quick overview of the
 presence of a given search term across archaeology-related tables/attributes 
 in the ROAD database and archaeology-related `road_get_*` functions.
It queries all relevant functions and returns a list indicating whether 
each function contains the search term. This helps you identify the appropriate 
`road_get_*` functions for your query.
Once identified, you can further query these functions to retrieve the 
relevant assemblages.
}
\details{
Background: The ROAD archaeology database is rich in categorized data, where 
archaeological findings are grouped according to their characteristics. These 
categories can be accessed using a variety of `road_get_*` functions, such as:

- `road_get_lithic_typology()`
- `road_get_lithic_raw_material()`
- `road_get_organic_tools()`
- `road_get_symbolic_artifacts()`
- `road_get_feature()`
- `road_get_miscellaneous_finds()`

Some search terms may not be easily categorized, while others may appear 
across multiple categories depending on contextual interpretations. This 
function assists in identifying the most relevant categories for your search.
}
\examples{
road_summarize_archaeology(term = "Cores")
}
