% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roadDB-package.R
\docType{package}
\name{roadDB-package}
\alias{roadDB}
\alias{roadDB-package}
\title{roadDB: Query data from the ROCEEH Out of Africa Database (ROAD)}
\description{
The aim of this package is to provide access to the ROCEEH Out of Africa Database (ROAD) and supply users with dataframes for further analysis in the R ecosystem.
}
\keyword{internal}
