% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_funcs.R
\name{set_database_path}
\alias{set_database_path}
\title{Set the package data directory path}
\usage{
set_database_path(path = NULL)
}
\arguments{
\item{path}{Character string. Path to the data directory.}
}
\description{
Set the package data directory path
}
